backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "Test-Scrapers"
authors = ["Aswin <aswin.subramanianmaheswaran@mongodb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
voyageai = "^0.3.2"
beautifulsoup4 = "^4.12.3" 
requests = "^2.31.0"        
feedparser = "^6.0.10"    
firecrawl-py = "^1.0.0"
tweepy = "^4.14.0"
praw = "^7.8.1"
nltk = "^3.9.1"
spacy = "^3.8.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
